uploading the web code #feel free to update it
A simple Web app for checking the duplication of 2 questions
Some Details - To run the file, first you have to download Streamlit, and then pyngrok, and then you have to go to the website. The first link that this code will give you, and then your website will be live.